backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.2
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
filelock==3.0.12
pbr==5.6.0
pipenv==2021.5.29
platformdirs==2.2.0
protobuf==3.17.3
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
stevedore==3.3.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
